requirements.txt
python-dotenv==1.0.0
PyMySQL==1.1.0
python-telegram-bot==20.6
peewee==3.17.0